requirements.txt
Django
djangorestframework
africastalking
phonenumbers
schedule
gunicorn
